Add files via upload
Python program that takes as input an excel file, generated by Orbit Intelligence, on the base of a patent search. The Excel file has two columns, one with the names of the Applicants (the long name of the column changed to "Applicants") and another with the Inventors. Applications with several applicants show the names separated by a new line \n, as well as with the names of the inventors.